Fix typings. Move Dropzone file back
