+++ package.json
@@ -1,7 +1,8 @@
 {
-  "name": "svelte-file-dropzone",
-  "version": "1.0.0",
+  "name": "arackaf-svelte-file-dropzone",
+  "version": "0.0.10",
   "description": "Svelte component for fileupload and file dropzone",
+  "main": "dist/index.js",
   "scripts": {
     "build": "rollup -c",
     "package": "svelte-kit sync && svelte-package && publint",
@@ -11,9 +12,14 @@
     "build-storybook": "build-storybook"
   },
   "exports": {
-		".": {
-			"types": "./dist/index.d.ts",
-			"svelte": "./dist/index.js"
+    "./Dropzone.svelte": {
+      "types": "./dist/components/Dropzone.svelte.d.ts",
+      "svelte": "./dist/components/Dropzone.svelte"
+    }
+	},
+  "typesVersions": {
+		">4.0": {
+			"Dropzone.svelte": ["./dist/components/Dropzone.svelte.d.ts"]
 		}
 	},
   "repository": {
@@ -62,6 +68,5 @@
     "dist"
   ],
   "svelte": "./dist/index.js",
-	"types": "./dist/index.d.ts",
-  "type": "module"
+	"types": "./dist/index.d.ts"
 }
