+++ src/lib/index.js
@@ -1,2 +1,2 @@
 // Reexport your entry components here
-export { Dropzone } from './Dropzone.svelte';
+export Dropzone from './components/Dropzone.svelte';
